rename folder using best practices
